Chore: UpdateBoard 타입분리
