+++ src/types/createPost.ts
@@ -17,3 +17,38 @@ export type MapInfoType = {
   content: string;
   position: { lat: number; lng: number };
 };
+export type UpdatePost = {
+  title: string;
+  content: string;
+  location: {
+    address?: string;
+    location: {
+      content: string;
+      position: {
+        lat: number;
+        lng: number;
+      };
+    };
+  };
+  categoryType: string;
+  eatTime: string;
+  alcohol: boolean;
+  joinLimit: number;
+  ageGroupLimit: boolean;
+  genderType: string;
+  linkUrl: string;
+  priceRange: number;
+};
+export type UpdatePostRequest = {
+  title: string;
+  content: string;
+  location: string;
+  categoryType: string;
+  eatTime: string;
+  alcohol: boolean;
+  joinLimit: number;
+  ageGroupLimit: boolean;
+  genderType: string;
+  linkUrl: string;
+  priceRange: number;
+};
